+++ src/imio/smartweb/core/contents/sections/video/content.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from imio.smartweb.common.config import DESCRIPTION_MAX_LENGTH
 from imio.smartweb.core.contents.sections.base import ISection
 from imio.smartweb.core.contents.sections.base import Section
 from imio.smartweb.locales import SmartwebMessageFactory as _
@@ -9,6 +10,16 @@ from zope import schema
 
 class ISectionVideo(ISection):
     """Marker interface and Dexterity Python Schema for SectionVideo"""
+
+    description = schema.Text(
+        title=_(u"Description"),
+        description=_(
+            u"Use **text** to set text in bold. Limited to ${max} characters.",
+            mapping={u"max": DESCRIPTION_MAX_LENGTH},
+        ),
+        max_length=DESCRIPTION_MAX_LENGTH,
+        required=False,
+    )
 
     video_url = schema.URI(
         title=_(u"Video url"),